Exception handling

+++ src/main/java/com/estudotiago/dscommerce/services/ProductService.java
@@ -1,9 +1,8 @@
 package com.estudotiago.dscommerce.services;
 
-import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -11,6 +10,12 @@ import org.springframework.stereotype.Service;
 import com.estudotiago.dscommerce.dto.ProductDTO;
 import com.estudotiago.dscommerce.entities.Product;
 import com.estudotiago.dscommerce.repositories.ProductRepository;
+import com.estudotiago.dscommerce.services.exception.DatabaseException;
+import com.estudotiago.dscommerce.services.exception.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
+
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -22,10 +27,11 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
 
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
-		ProductDTO dto = new ProductDTO(product);
-		return dto;
+		
+		Product product = repository.findById(id).orElseThrow(() -> 
+		new ResourceNotFoundException("Recurso não encontrado"));
+		return new ProductDTO(product);
+		
 	}
 
 	@Transactional(readOnly = true)
@@ -50,19 +56,31 @@ public class ProductService {
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
 
+		try {
 		Product entity = repository.getReferenceById(id);
 		copyDtoEntity(dto, entity);
 		entity = repository.save(entity);
 
 		return new ProductDTO(entity);
-
+		
+		}catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Rescurso não encontrado");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
+
 
 	private void copyDtoEntity(ProductDTO dto, Product entity) {
 		entity.setName(dto.getName());
